+++ examples/node.js
@@ -1,5 +1,0 @@
-var api = require('../src/api');
-
-api.parse(__dirname + '/stylesheets').then(function (results) {
-  console.log(results);
-});